Moved to Java 17

+++ io-hotmoka-tests/src/test/java/io/hotmoka/tests/Signatures.java
@@ -59,7 +59,7 @@ class Signatures extends HotmokaTest {
 
 	@Test @DisplayName("create accounts with distinct signing algorithms")
 	void createAccountsWithDistinctSigningAlgorithms() throws TransactionException, CodeExecutionException, TransactionRejectedException, InvalidKeyException, SignatureException, NoSuchAlgorithmException {
-		var amount = new IntValue(_10_000_000);
+		var amount = new IntValue(_10_000_000.intValue());
 
 		var sha256dsa = SignatureAlgorithms.sha256dsa();
 		KeyPair sha256dsaKeyPair = sha256dsa.getKeyPair();
@@ -79,7 +79,7 @@ class Signatures extends HotmokaTest {
 
 	@Test @DisplayName("create accounts with distinct signing algorithms and use them for signing transactions")
 	void createAccountsWithDistinctSigningAlgorithmsAndUseThem() throws TransactionException, CodeExecutionException, TransactionRejectedException, InvalidKeyException, SignatureException, NoSuchAlgorithmException {
-		var amount = new IntValue(_10_000_000);
+		var amount = new IntValue(_10_000_000.intValue());
 		var callee = new NonVoidMethodSignature("io.takamaka.code.lang.Coin", "panarea", ClassType.BIG_INTEGER, BasicTypes.LONG);
 
 		var sha256dsa = SignatureAlgorithms.sha256dsa();
